requirements.txt
streamlit
snowflake-connector-python
openai
python-dotenv
flask
